use SpillManager in benchmark
