+++ datafusion/physical-plan/src/lib.rs
@@ -50,6 +50,7 @@ pub use crate::ordering::InputOrderMode;
 pub use crate::stream::EmptyRecordBatchStream;
 pub use crate::topk::TopK;
 pub use crate::visitor::{accept, visit_execution_plan, ExecutionPlanVisitor};
+pub use spill::spill_manager::SpillManager;
 
 mod ordering;
 mod render_tree;
